+++ index.html
@@ -126,6 +126,6 @@
       </div>
     </section>
   </main>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
